Enhance PostgreSQL extensions installer with testing capabilities and version detection
- Updated GitHub Actions workflow to include testing step.
- Added support for detecting PostgreSQL version from the base image.
- Introduced CLI arguments for test mode and PostgreSQL version.
- Enhanced Dockerfile to conditionally run tests based on TEST_MODE.
- Updated README files with new status badges and improved instructions.
- Added configuration management for tests.
