+++ Cargo.toml
@@ -5,4 +5,5 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+clap = { version = "4.5.50", features = ["derive"] }
 tokio = { version = "1.48.0", features = ["full"] }
